Proposed classes re-locations
